Finished python script. Tests needed.
The user is registered and logged in with create() and login(). Then he creates an event and enrolls himself with create_event() and enroll(). Then he sends location data to the event API with send_loc(). Need to still write tests for this and looping it for 100 people at the same time, will most likely use multiprocessing module which supports concurrency and avoids python's GIL in order to act like 100 people. Might be too resource hungry, but worth a shot.